+++ web-client/app/styles/fonts.css
@@ -3,5 +3,5 @@
   font-style: normal;
   font-weight: 400;
   font-display: swap;
-  src: url("../public/NotoSans-Regular.ttf") format("truetype");
+  src: url("../../public/NotoSans-Regular.ttf") format("truetype");
 }
